# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FileStore API
  version: 0.1.0
  description: |
    Contract for the FileStore MVP. Health operations are implemented in stage 0;
    product operations marked as planned are implemented in subsequent stages.
...
servers:
  - url: /
    description: Same origin as the API endpoint
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered user and an authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
  /auth/login:
    post:
      tags: [Authentication]
      summary: Log in with email and password
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication token and current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the current authentication token
      operationId: logoutUser
      responses:
        "204":
          description: Token revoked
        "401":
          $ref: "#/components/responses/ProblemResponse"
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get the authenticated user
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/ProblemResponse"
  /workspaces/base:
    get:
      tags: [Workspaces]
      summary: Get the reserved base workspace
      operationId: getBaseWorkspace
      responses:
        "200":
          description: Base workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/ProblemResponse"
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces available to the current user
      operationId: listWorkspaces
      responses:
        "200":
          description: Workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/ProblemResponse"
    post:
      tags: [Workspaces]
      summary: Create a private workspace
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
  /workspaces/{workspaceId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
    get:
      tags: [Workspaces]
      summary: Get a workspace
      operationId: getWorkspace
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
  /workspaces/{workspaceId}/members:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
    post:
      tags: [Workspaces]
      summary: Add or update a workspace member by email
      operationId: putWorkspaceMemberByEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberByEmailRequest"
      responses:
        "200":
          description: Workspace member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMember"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "403":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
  /workspaces/{workspaceId}/members/{userId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
      - $ref: "#/components/parameters/UserID"
    put:
      tags: [Workspaces]
      summary: Add or update a workspace member
      operationId: putWorkspaceMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRoleRequest"
      responses:
        "200":
          description: Workspace member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMember"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "403":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
    delete:
      tags: [Workspaces]
      summary: Remove a workspace member
      operationId: removeWorkspaceMember
      responses:
        "204":
          description: Member removed
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "403":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
  /workspaces/{workspaceId}/files:
    parameters:
      - $ref: "#/components/parameters/WorkspaceID"
    get:
      tags: [Files]
      summary: List files in a workspace
      description: Cursor order is case-folded file name ascending, then file ID ascending.
      operationId: listFiles
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageLimit"
      responses:
        "200":
          description: Files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilePage"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
    post:
      tags: [Files]
      summary: Upload a file and create version 1
      operationId: createFile
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [content]
              properties:
                name:
                  type: string
                  minLength: 1
                text_encoding:
                  type: string
                  default: utf-8
                  enum: [utf-8, utf-16le, utf-16be, windows-1251]
                content:
                  type: string
                  format: binary
                  description: Streamed content with a default hard limit of 104857600 bytes.
      responses:
        "201":
          description: Created file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "403":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
        "413":
          $ref: "#/components/responses/ProblemResponse"
        "423":
          $ref: "#/components/responses/ProblemResponse"
  /files/{fileId}:
    parameters:
      - $ref: "#/components/parameters/FileID"
    get:
      tags: [Files]
      summary: Get file metadata
      operationId: getFile
      x-implementation-status: implemented
      responses:
        "200":
          description: File metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
...
  /files/{fileId}/versions:
    parameters:
      - $ref: "#/components/parameters/FileID"
    get:
      tags: [Files]
      summary: List published file versions
      description: Cursor order is version number descending.
      operationId: listFileVersions
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageLimit"
      responses:
        "200":
          description: Versions ordered by version number descending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileVersionPage"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
  /files/{fileId}/content:
    parameters:
      - $ref: "#/components/parameters/FileID"
      - $ref: "#/components/parameters/VersionNumber"
    get:
      tags: [Files]
      summary: Download current or selected file bytes
      operationId: downloadFile
      x-implementation-status: implemented
      responses:
        "200":
          description: Immutable file bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
...
    ProblemResponse:
      description: Request failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 12
          writeOnly: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          writeOnly: true
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          minLength: 43
          maxLength: 43
          pattern: "^[A-Za-z0-9_-]{43}$"
          description: Opaque bearer token returned only by register/login and stored only as a hash by the API.
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, is_superadmin, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        is_superadmin:
          type: boolean
        created_at:
          type: string
          format: date-time
...